Create two constructors
The data typically lives in a csv, both the text and the labels.
This makes using initialization lists tricky as we initialize two
variables within the read_data function, but return them as a
tuple. So we need to use two initialization lists, one public, and
one private, in order to read in the data and then parse the tuple.